Added: Read config file .chaas.yml. Refactored structure.
